tests/wpt/web-platform-tests/css/css-page/parsing/page-valid.html

<!DOCTYPE html>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://www.w3.org/TR/css-page-3/#using-named-pages">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  test_valid_value("page", "auto");
  test_valid_value("page", "table");
  test_valid_value("page", "xyzabc");
</script>
